requirements.txt
Django==1.8
djangorestframework==3.1.1
Unipath==1.1
